change default avatar to identicon instead of 404 (#618)
When a user logs in, the server attempts to display their gravatar image in the navbar. Currently, if the image for that user cannot be found, it defaults to "404". The 404 default is why you will see the browsers 404 image placeholder momentarily, which varies depending on the browser. The commit changes the default image to an identicon to resolve this. There is already a stimulus controller set up to handle this default, but the image is rendered faster than the javascript is able to replace the image with the fallback. This is why you will see the 404 image briefly, which is then changed to the identicon. Since we are now defaulting the gravatar image to be an identicon, there is no reason to have the fallback-image javascript run here, and it has been removed.

Fixes #378