update ngrok setup (#677)
Because of ActionDispatch::HostAuthorization, we cannot dynamically start an ngrok tunnel while the app is booting because the ngrok host is not in the environment, and thus not seen as an accepted host. I have not found a way to insert the generated ngrok subdomain into config.hosts bc it is middleware and you have to restart the app to reload it, which causes the tunnel subdomain to change.