fix invitations form (#1084)
Fixes issues currently causing this form to not work. If you do not select an organization, which is not required, then the form will not submit, and you will not be taken back to the correct page. This also fixes some issues with error messages not showing up on the form, and layout issues with the invitations controller.