fix issues with blank fields on campaign form (#1099)
The creatives issue was just caching but the issue with the user id not showing up seemed to be because of the sync to campaign bundle method. I don't see a reason why we need to sync this field, especially since the person who creates the bundle will be eric for the near future.