move all images from sprockets to webpacker (#1083)
Move all of images out of the legacy assets folder and under webpackers control.

It would be good to eventually remove all dependency on sprockets. The other benefit of this change is it should dramatically drop our slug size on Heroku since I was able to remove all of the images we aren't using.